requirements.txt
python-dotenv==1.0.0
anthropic==0.7.0
pandas==2.2.1
numpy==1.26.4
openai==1.3.0
google-auth-oauthlib==1.0.0
google-auth==2.22.0
google-api-python-client==2.95.0
tenacity==8.2.2
asyncio==3.4.3
requests==2.31.0
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1
